app/requirements.txt
python_binance
flask
jsonify
pandas